[None][refactor] Align dynamic tree verify kernel dtype to int32
Align verify_dynamic_tree_greedy_packed kernel and its Python callers to
use int32 end-to-end, matching TRT-LLM's main framework token dtype
convention (accepted_tokens / next_draft_tokens are int32).

Changes:
- Kernel header/body: all verify params (predicts/acceptIndex/
  acceptTokenNum/acceptToken/candidates/retrievePacked/targetPredict)
  switched from int64_t to int32_t
- thop dispatcher: TORCH_CHECK(kInt32) and data_ptr<int32_t>() for all
  verify tensors
- Python side: dynamic_tree_ops.py allocates int32 buffers directly;
  eagle3_dynamic_tree.py drops redundant .to(torch.int32) casts

parentList and selectedIndex in build kernel remain int64 (they are
indices into arrays, not token IDs).

Validated by 4-test regression on H200 (worktree build, rc13):
  LLaMA-8B  d60  AR=4.508  (min 4.50)
  LLaMA-8B  d30  AR=4.215  (min 4.20)
  Qwen-235B d30  AR=2.919  (> baseline 2.75)
  Qwen-235B d60  AR=3.116  (> baseline 2.88)

+++ cpp/tensorrt_llm/kernels/speculativeDecoding/dynamicTreeKernels.cu
@@ -266,8 +266,8 @@ void invokeBuildDynamicTree(int64_t const* parentList, int64_t const* selectedIn
 
 //! retrievePacked layout [bs, numDraftTokens, 3] int32 row-major:
 //! [b,n,0]=retrieveIndex, [b,n,1]=retrieveNextToken, [b,n,2]=retrieveNextSibling
-__global__ void verifyDynamicTreeGreedyPackedKernel(int64_t* predicts, int64_t* acceptIndex, int64_t* acceptTokenNum,
-    int64_t* acceptToken, int64_t const* candidates, int32_t const* retrievePacked, int64_t const* targetPredict,
+__global__ void verifyDynamicTreeGreedyPackedKernel(int32_t* predicts, int32_t* acceptIndex, int32_t* acceptTokenNum,
+    int32_t* acceptToken, int32_t const* candidates, int32_t const* retrievePacked, int32_t const* targetPredict,
     bool const* treeValid, uint32_t batchSize, uint32_t numSpeculativeTokens, uint32_t numDraftTokens)
 {
     uint32_t bx = blockIdx.x;
@@ -297,8 +297,8 @@ __global__ void verifyDynamicTreeGreedyPackedKernel(int64_t* predicts, int64_t* 
         while (curIndex >= 0 && static_cast<uint32_t>(curIndex) < numDraftTokens)
         {
             int32_t draftLocalIdx = row[static_cast<size_t>(curIndex) * 3];
-            int64_t draftTokenId = candidates[batchOffset + curIndex];
-            int64_t targetTokenId = targetPredict[batchOffset + lastAcceptedLocalIdx];
+            int32_t draftTokenId = candidates[batchOffset + curIndex];
+            int32_t targetTokenId = targetPredict[batchOffset + lastAcceptedLocalIdx];
 
             if (draftTokenId == targetTokenId)
             {
@@ -323,8 +323,8 @@ __global__ void verifyDynamicTreeGreedyPackedKernel(int64_t* predicts, int64_t* 
     predicts[batchOffset + lastAcceptedLocalIdx] = targetPredict[batchOffset + lastAcceptedLocalIdx];
 }
 
-void invokeVerifyDynamicTreeGreedyPacked(int64_t* predicts, int64_t* acceptIndex, int64_t* acceptTokenNum,
-    int64_t* acceptToken, int64_t const* candidates, int32_t const* retrievePacked, int64_t const* targetPredict,
+void invokeVerifyDynamicTreeGreedyPacked(int32_t* predicts, int32_t* acceptIndex, int32_t* acceptTokenNum,
+    int32_t* acceptToken, int32_t const* candidates, int32_t const* retrievePacked, int32_t const* targetPredict,
     bool const* treeValid, SizeType32 batchSize, SizeType32 numDraftTokens, SizeType32 numSpecStep, cudaStream_t stream)
 {
     dim3 grid(batchSize);
